treeproof: build merkle trees from depth so huge list limits don't panic
GetTree previously panicked on a list whose ssz-max exceeds the platform
int range (e.g. the practically-unbounded 18446744073709551615), while
HashTreeRoot computed the root fine. The wrapper narrowed the uint64
limit to int and the tree builder rounded it up with int arithmetic,
overflowing to a negative value.

A list capacity only determines the tree depth (at most 64), so tree
construction is now driven by that depth via chunkLimitDepth (mirroring
the hasher), keeping it overflow-free for any capacity. The public
TreeFromNodes/TreeFromNodesWithMixin gain uint64 forms (TreeFromNodes64,
TreeFromNodesWithMixin64) that carry the canonical logic; the int forms
forward with per-platform range checks, so the API stays compatible.

The Wrapper's tree-building machinery (Commit*, AddEmpty) is unexported:
it is not part of the HashWalker contract and was only ever used
internally. The MerkleizeWithMixin length/limit range guards are gone
since the values are now handled correctly.
